Move controller tests to V0 subdir.
